Building the puzzle scene
So far in theory, the board can be built, pieces can be built and added to the shop.
All of this is testing pending, and numbers for locations were pulled out of thin air.
Aside from testing, adding values to pieces would be good, allowing pieces to be put onto the board, allow the board to detect them, and if I have time rotation.
Rotation is going to suck because the pieces are just arrays of squares, so going to have to make a modified matrix transform function. :)